Child classes for abstract FiniteStateMachine depending on StorageUnit
