+++ src/main/java/org/icatproject/ids/thread/DfArchiver.java
@@ -7,12 +7,11 @@ import java.util.List;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import org.icatproject.ids.FiniteStateMachine;
 import org.icatproject.ids.LockManager.Lock;
 import org.icatproject.ids.PropertyHandler;
 import org.icatproject.ids.plugin.DfInfo;
 import org.icatproject.ids.plugin.MainStorageInterface;
+import org.icatproject.ids.v3.FiniteStateMachine.FiniteStateMachine;
 
 /*
  * Removes datafiles from the fast storage (doesn't write them to archive storage)
