+++ src/main/java/org/icatproject/ids/thread/DsRestorer.java
@@ -16,7 +16,6 @@ import org.slf4j.LoggerFactory;
 
 import org.icatproject.Datafile;
 import org.icatproject.Dataset;
-import org.icatproject.ids.FiniteStateMachine;
 import org.icatproject.ids.IcatReader;
 import org.icatproject.ids.IdsBean;
 import org.icatproject.ids.LockManager.Lock;
@@ -25,6 +24,7 @@ import org.icatproject.ids.plugin.ArchiveStorageInterface;
 import org.icatproject.ids.plugin.DsInfo;
 import org.icatproject.ids.plugin.MainStorageInterface;
 import org.icatproject.ids.plugin.ZipMapperInterface;
+import org.icatproject.ids.v3.FiniteStateMachine.FiniteStateMachine;
 
 /*
  * Restores datafiles from the slow to the fast storage.
