+++ src/main/java/org/icatproject/ids/thread/DsWriter.java
@@ -15,7 +15,6 @@ import org.slf4j.LoggerFactory;
 import org.icatproject.Datafile;
 import org.icatproject.Dataset;
 import org.icatproject.ids.DfInfoImpl;
-import org.icatproject.ids.FiniteStateMachine;
 import org.icatproject.ids.IcatReader;
 import org.icatproject.ids.IdsBean;
 import org.icatproject.ids.LockManager.Lock;
@@ -24,6 +23,7 @@ import org.icatproject.ids.plugin.ArchiveStorageInterface;
 import org.icatproject.ids.plugin.DsInfo;
 import org.icatproject.ids.plugin.MainStorageInterface;
 import org.icatproject.ids.plugin.ZipMapperInterface;
+import org.icatproject.ids.v3.FiniteStateMachine.FiniteStateMachine;
 
 /**
  * Copies dataset from main to archive
