+++ src/main/java/org/icatproject/ids/v3/FiniteStateMachine/FiniteStateMachineForSingleLevelStorage.java
@@ -1,0 +1,27 @@
+package org.icatproject.ids.v3.FiniteStateMachine;
+
+import org.icatproject.ids.IcatReader;
+import org.icatproject.ids.LockManager;
+import org.icatproject.ids.exceptions.InternalException;
+import org.icatproject.ids.v3.enums.DeferredOp;
+import org.icatproject.ids.v3.models.DataInfoBase;
+
+public class FiniteStateMachineForSingleLevelStorage extends FiniteStateMachine {
+
+    protected FiniteStateMachineForSingleLevelStorage(IcatReader reader, LockManager lockManager) {
+        super(reader, lockManager);
+    }
+
+
+    @Override
+    public void queue(DataInfoBase dataInfo, DeferredOp deferredOp) throws InternalException {
+        throw new InternalException("### Operation is not permitted for single level storage");
+    }
+
+
+    @Override
+    protected void scheduleTimer() {
+        //nothing to do here for single level storage
+    }
+
+}
