+++ src/main/java/org/icatproject/ids/v3/ServiceProvider.java
@@ -3,13 +3,13 @@ package org.icatproject.ids.v3;
 import java.util.Set;
 
 import org.icatproject.ICAT;
-import org.icatproject.ids.FiniteStateMachine;
 import org.icatproject.ids.IcatReader;
 import org.icatproject.ids.LockManager;
 import org.icatproject.ids.PropertyHandler;
 import org.icatproject.ids.Transmitter;
 import org.icatproject.ids.exceptions.InternalException;
 import org.icatproject.ids.plugin.MainStorageInterface;
+import org.icatproject.ids.v3.FiniteStateMachine.FiniteStateMachine;
 import org.icatproject.ids.v3.enums.CallType;
 
 /**
