+++ src/main/java/org/icatproject/ids/v3/UnfinishedWorkServiceBase.java
@@ -10,10 +10,10 @@ import org.icatproject.Datafile;
 import org.icatproject.Dataset;
 import org.icatproject.IcatExceptionType;
 import org.icatproject.IcatException_Exception;
-import org.icatproject.ids.DeferredOp;
 import org.icatproject.ids.StorageUnit;
 import org.icatproject.ids.exceptions.InsufficientPrivilegesException;
 import org.icatproject.ids.exceptions.InternalException;
+import org.icatproject.ids.v3.enums.DeferredOp;
 import org.icatproject.ids.v3.models.DataFileInfo;
 import org.icatproject.ids.v3.models.DataSetInfo;
 import org.icatproject.utils.IcatSecurity;
